17/03/2025 1:35AM: -Minor changes, merge other commit files due to error in a pull-process

+++ src/main/java/com/se1873/js/springboot/project/dto/RequestCreationRequestDTO.java
@@ -10,7 +10,7 @@ import lombok.Setter;
 @AllArgsConstructor
 @NoArgsConstructor
 public class RequestCreationRequestDTO {
-    private Long departmentId;
+    private Integer departmentId;
     private String requestType;
     private String employeeName;
     private Double salaryIncreasePercentage;
